+++ autofocus/src/main/java/org/micromanager/autofocus/OughtaFocus.java
@@ -39,8 +39,8 @@ import mmcorej.StrVector;
 import mmcorej.TaggedImage;
 import org.micromanager.AutofocusPlugin;
 import org.micromanager.Studio;
-import org.micromanager.autofocus.internal.BrentFocuser;
-import org.micromanager.autofocus.internal.ImgSharpnessAnalysis;
+import org.micromanager.autofocus.internal.oughtafocus.BrentFocuser;
+import org.micromanager.autofocus.internal.oughtafocus.ImgSharpnessAnalysis;
 import org.micromanager.internal.utils.AutofocusBase;
 import org.micromanager.internal.utils.imageanalysis.ImageUtils;
 import org.micromanager.internal.utils.MMException;
